SideBySide2: Fix max width of comment widget groups
The group should have a max-width of 650px. This allows the group
to show the right side drop shadow on a normally sized window,
making the comment appear to float a bit on the page.

The max width was incorrectly applied to the header line of the
comment, andwas not being used by the browser to size the group.

Change-Id: I564c7fd9b154d64cdf1b20836351f40954a4900d

Former-commit-id: 763c5df979ce81bfec6894ff51b7b13c7e1e8b24