SideBySide2: Reclaim more browser keys from vim
Ctrl-R is reload page. This is useful when developing a web UI.
In vim mode it means redo a change which was undone, which is
useless on a read-only buffer. Delete Ctrl-R and allow the browser
to handle it.

Also delete Ctrl-P (previous line, useless) and Ctrl-S.

Delete Ctrl-{0..9} so Chrome can switch between tabs.

Change-Id: I22da04bcf7d9754787214c1b7321e5ddfaf8520b
Former-commit-id: 4e73c9330292523e8a3542d9aa9f653a0b36af3b