Simplify documentation indexer
Write the index entirely in RAM. Its only a few hundred KiB, which
trivially fits in memory.

Compress it twice in memory, once to build the ZIP that is unpacked
at runtime, and again to package it into a JAR for linking with
the runtime. This saves a build step in the BUCK rules.

Move the ZIP under the server package name, to reduce any risk
of collision with another concept of "index.zip".

Change-Id: I74e59712e9855ac79c5220ff0a6b30ecbc3d152f

Former-commit-id: a42c0bb09a50f881634ebd14b90cc4c453d863c6