Clean left over data migration after removal of TrackingIds table
2aac74a24dfc39f0ea580da510d71394e6dac38f [formerly 9f4de526dd3fec0400f6df53628956a4bf0d9033] removed TrackingIds table.
35d68e988fef80db5a1fc06b820bf13104d1e9d1 [formerly 766e4415b0f004a4e53c93d256624361072c67c4] tries to swap primary key from
tracking_id to tracking_key in data migration method.

But because the table was removed from the Schema the TrackingIds is not
picked by gwtorm for the upgrade (ALTER TABLE add tracking_key statement).

In Schema_70 the migration step tries to execute:

  UPDATE tracking_ids SET tracking_key = tracking_id

and is failing because the column tracking_key wasn't created.

To rectify this, remove the migrate data method entirely in Schema_70.

Bug: Issue 2382
Change-Id: I1d7b6deb50be8595626d210a4350683d8b996de5

Former-commit-id: 2d6629742c11e1a5d03f7a7caf939d1d3748547a