Plugin API: Add datasource interception extension point
This change extends the plugin API to support monitoring of Gerrit's
SQL activity by providing the extension point to create a DataSource
proxy to intercept execution of SQL statements.

To activate SQL monitoring the following must be done:

Add the interceptor class to gerrit.config:

  dataSourceInterceptorClass = javamelody.JavamelodyInterceptor

Implement the interface DataSourceInterceptor:

  public class JavamelodyInterceptor implements DataSourceInterceptor {
    @Override
    public DataSource intercept(String name, DataSource dataSource) {
      return JdbcWrapper.SINGLETON.createDataSourceProxy(name, dataSource);
    }
  }

Put the jar containing the interceptor into $gerrit_site/lib directory.

Working example of this concept can be found in javamelody-plugin.

Change-Id: I5f0b5ffe072394eae77e3b1735a76dceff4bdc42
Former-commit-id: 38a6f6cc2d548a8017c5b7c0f273a53d1b0e8184