Revert "Scroll from the window when rendering entire file"
This reverts commit f6a42795e39c2c9f4fa077d663b049e4b88ee3d8 [formerly b07a26f2ebb24d1e9a9a20c7683fed55aab9a6bd].
Codemirror expects to be able to scroll Codemirror instance in
order to execute any line navigation. Furthermore, horizontal
scrollbars are rendered in the Codemirror at the bottom of the
document (not window). In order to have a "window" scroll bar,
we would likely have to steal the scroll bars inside Codemirror
and make it look like the window scroll bar.

Change-Id: I3199edc039ba716723eaf8032eaa5355c2100666

Former-commit-id: d29c67d4a26a88fd7df52e605ee42be8ba6a8cb5