Remove quotes and trailing period from "topic edited" messages
When the topic is edited, a change message is added that shows
the old and new topics.  If the topic is set to a URL, the URL
is linkified, but the linkification includes the trailing closing
quote and period after the topic name, and the link does not work.

Remove the quotes and trailing period to prevent this.

Bug: Issue 2397
Change-Id: Ic70882f7355cf84e12a2853f0ac6a77c4ca1543d

Former-commit-id: f97dbd8e7bb00cbaabc4ee48a93b89a92b7f0759