Remove query limit when drilling down from a dashboard to a query
By clicking on the section header in a custom dashboard one can
navigate to the query of that section. In a dashboard space is limited
which is why the section queries often have a limit operator so that a
section cannot grow too big. Sometimes it makes sense to drill down
into a query and look at the complete query result. Remove any limit
operator automatically when drilling down into a section query.

Change-Id: Ic346f66fb7dec3911e9798f03f9301e87af4d7c0
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: b0f9afa0fdd1307be77f7a090e8e957003df3613