Remove limited patch set view of ChangeData
This required extra logic and state and complicated the semantics of a
ChangeData. It was only used by ChangeJson to limit results to a
single patch set. Instead of supporting arbitrary subsets of
revisions, just have ChangeJson do the limiting itself. The only
additional fetching this might require is for a single non-current
patch set, we might fetch all patch set rows for the change from the
db.

Change-Id: I7a35369d11da3b4b4d7f134766991f33c4a0b65c

Former-commit-id: e759cd16ab45578a7182a1c0490c2f26955da0f8