Add link to validation listeners in the plugin documentation
The plugin documentation describes listeners that plugins can
implement to be notified about certain operations in Gerrit.

The validation listeners are described on an own page, but readers of
the plugin documentation may expect the description of the validation
listeners to be next to the description of the other listeners.

Change-Id: Id12ba70f00b6f08eed14d6c4644de50f4f2657d8
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 32737600d7ed70a26970211f37dcada3096b9219