Remove obsolete commands from slave mode commands list
Running the `gerrit` command without any arguments over ssh results
in a list of available commands being shown.

When running the Gerrit server in slave mode, this list includes the
`approve` command which was deprected and then removed, and the
`replicate` command which is now handled by the replication plugin.

Remove these commands from the list.

Change-Id: Ide32dde34dcb1bc59a05001c43d2ce423d193ba0

Former-commit-id: 80b4c8f111074b21b5115ad0d7d694a0a7cbf7fb