Refactor the way of binding ssh commands
If a command should only be used on a server in master mode, but not
slave mode, it should be bound in both MasterCommandModule and
SlaveCommandModule, we have to remember to add it separately on each
of the commands list, the maintenance work is tedious.

Refactor the way of binding ssh commands. Add a field 'runsWith' to
CommandMetaData to mark a given ssh command to run on MASTER mode or
MASTER_OR_SLAVE mode. Remove SlaveCommandModule and MasterCommandModule,
and use only DefaultCommandModule.

This reverts commit 0b235a32f50bd3e12d933e3b22ea10e932b03ba6 [formerly 84bcead690240e07e0314b165fb8d6dc95d2c5e7].

Change-Id: Ia4908267b135be04ff9c8a9470a243a6c8d665a8

Former-commit-id: ae479c7fda95a65e9132e0eb808284789fb9a7c4