Don't allow project owners to create branches if create is blocked
For creating a branch by pushing to Gerrit it is required to have the
create reference permission. For convenience project owners are
allowed to create branches through WebUI/REST/SSH without create
reference permission (because they anyway can assign this permission
themselves). However create reference can be blocked on a parent
project to prevent project owners from creating branches. In this case
assigning create reference on a project has no effect. The problem is
that project owners can still create the branches through
WebUI/REST/SSH although create reference is actually blocked for them.
With this change project owners are only allowed to create branches if
create reference is not blocked for them. Administrators can still
create branches even if create reference is blocked.

This is consistent to how permissions for branch deletions are
handled, see
  https://gerrit-review.googlesource.com/50890

Change-Id: I1c94aade65e36282400d9d0d407fcc78bbd3a2ae
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 6e43e2aefaf03c56e2fdf99f9077949d38604939