Allow plugins to check for config updates when reading plugin config
Plugins can have their global configuration in a plugin subsection of
the gerrit.config file. PluginConfigFactory provides easy access to
this global plugin configuration and ensures that the gerrit.config
file is not on any access loaded from the filesystem. The
GerritServerConfig is injected into the PluginConfigFactory
so that updates of the gerrit.config file only get visible in
PluginConfigFactory after the server has been restarted.

Plugins may want to provide ways to live update their global
configuration, e.g. they could provide a REST endpoint to update
their plugin section in gerrit.config. To make the changes visible
in PluginConfigFactory the cached GerritServerConfig in it needs
to be reloaded. For this purpose this change adds a refresh flag
to the method in PluginConfigFactory that reads the global plugin
config. The GerritServerConfig is then reloaded when it is was
changed in the filesystem.

Change-Id: Ide6b1457bfe5a63459e816ca9e00604a62da8365
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 9ed9540325b9bcd2f6566d1e0de5a7fec781335a