Acceptance tests: Disable HTTPD per class or method annotation
If HTTPD not used in tests, it doesn't make much sense to enable it.

Using @NoHttpd on class or method it is now possible to disable HTTP
entirely and is another small step making slow tests fast.

With maturity of plugin API, more and more tests are going to move
to plugin API driven tests from REST driven tests.

Benchmark shows, that doing REST vs. plugin API comparisson:

  @Test
  public void rest() throws Exception {
    String changeId = createChange().getChangeId();
    for (int i = 0; i < 1000; i++) {
      getRESTChange(changeId, ALL);
    }
  }

  @Test
  public void api() throws Exception {
    String changeId = createChange().getChangeId();
    for (int i = 0; i < 1000; i++) {
      getPluginAPIChange(changeId, ALL);
    }
  }

make on my machine ca. 2 sec. for plugin API vs. 10 sec. for REST:
The whole Servlet/HTTP stack and Gson serialization
and deserialization is out if the picture for API driven tests.

Change-Id: If60170f06f466765d9602450043a1629a344d7ce

Former-commit-id: 83ea0403b118c99a9986d1a7fc46247c644ea4b2