Run acceptance tests with custom suite to set Configs
@Parameterized doesn't allow for running a single test method in
Eclipse, since Eclipse looks for a test name exactly matching the
method name, and @Parameterized always appends "[name]" to method
names.

Use a custom suite that avoids this problem by not appending anything
to the name for the default config. Running a single test method in
Eclipse now chooses the default (empty) config. This is arguably not
as nice as if it ran all permutations with that name, but it at least
chooses the least surprising option.

Since this suite only deals with setting a single field of a single
type on the test, it can be simpler than the more general
implementation of @Parameterized. It also has the advantage that
subclasses can provide their own @ConfigSuite.Config implementations
to add additional permutations at the per-acceptance-test level,
rather than globally.

Change-Id: I2a83f789f922d894c1bd111c1339f7ca9aacab00

Former-commit-id: 9760cdc2b2aff18194724b4f1077fb6ada811692