Don't validate LabelTypes in ChangeUpdate
We don't perform other types of validation on the approvals input,
e.g. for permissions, which are just handled by LabelNormalizer. Since
it's assumed callers will be passing their inputs through
LabelNormalizer if necessary anyway, they are likely already taking
care of validation.

Still pass in the Comparator for aesthetics when ordering (analogous
to how we use account names for aesthetics when referring to
identities).

Change-Id: I1e8005db9cc867a60cfb384bbec55e363f250dac

Former-commit-id: 9f77aede3d02639e762ba2623095983fbc83af4d