Improve information message about rebuilding the index
When initialising Gerrit into an existing workspace, the index
should be "rebuilt", not "built".  Change the message accordingly.

Also add a new line at the beginning of the message to make it
stand out from the other information that is being output.  This
is consistent with other warnings such as the warning about
bouncycastle not being bundled.

Change-Id: I2ef5b9e141097da70fafd9dcf266988e9d6aaa27

Former-commit-id: 331dd5f199d2ce8dd6b0c9f6a2f02035679e0ede