Solr index URL should be configured under the main "index" section
Configuration of the Solr index URL must currently be done in a
subsection, as follows:

  [index]
    type = SOLR
  [index "solr"]
    url = 127.0.0.1

Subsections are also being used for configuration of specific indexes
like the "changes_open" and "changes_closed" subsections that are used
for Lucene indexes, and are documented as "index.name.xyz".

Having the Solr URL configration under a similar subsection may cause
confusion, and it would be more intuitive to configure the URL under
the main index section, like:

  [index]
    type = SOLR
    url = 127.0.0.1

Subsections are being used for configuration of specific indexes like
the "changes_open" and "changes_closed" settings that can be specified
for Lucene indexes which are documented as "index.name.xyz"

Also add missing documentation of the URL parameter.

Change-Id: I0bd9eabbdbfa29241688c738aa5d213e726c90b0

Former-commit-id: ac88c36c5da6db2ef5daa14c793f0f6eecfc41fa