Handle exceptions during reindexing
If an exception is thrown from within the change indexer, the reindex
program shows a stack trace.  In the case of an error when using the
Solr index, it does not exit and continues running, while showing debug
messages "Got ping response for sessionid", but not doing anything
else.

Catch exceptions in the indexer and exit with a fatal error message.

Change-Id: Ief41eccc0d3fe9c5592782b51afa66c0c3150886

Former-commit-id: 6ab111e94f71fa5b2d63c45df62cc9006dde3ef6