Control which plugins will be auto installed on auto init
Use the system property 'gerrit.install_plugins' to control which
of the packaged plugins will be installed during auto site init
in a servlet container.

When 'gerrit.install_plugins' property is not defined then all packaged
plugins will be installed. If it is present then it is parsed as a
comma separated list of plugin names to install. Defining an empty
string as the property value means that no plugin will be installed.

Change-Id: I52881c9677bc9cc1e51c46c2c2ce4558c79b9a3e

Former-commit-id: 9f058b37bc5626b7a6d1d28383d9f531bf0bd96f