Plugin support for revision weblink
Support for an extension point that allows for weblinks to
several external tools.

First effort, weblinks for patch sets.

At the present time GitWebLink is integrated. But the aim is to
extend support to engulf entire GitWebLink and provide backwards
support by adding a plugin that has the same functionality as
GitWebLink has today.

Change-Id: I19f4d00325d2991df9514ee8833a8873649c27bc

Former-commit-id: ae1a10ce48850a9527f08a5d476d0c4cc41a75e8