Allow submit on behalf of any user regardless of visibility
To allow a CI account to submit on behalf of another user, an
administrator must already grant a special permission to that
account that allows bypassing submit restrictions. Also allow this to
bypass the accounts.visibility restriction.

Do enforce that the target account be able to see the destination ref.

Change-Id: I4ab1962c4d415a4eff3b941b3c1ae935cc56bbd2

Former-commit-id: 894f136df8ad5c1cf0efd243ba186fd55dd68b15