Allow creation of persistent caches after server is started
98802547bd0e72fbf48f34d3c3cc98c91eabacc1 [formerly 2e1cb2b8495761a7e9abc09ebca24ca6bbf2afbe] switches from Ehcache to Guava
for providing caching in Gerrit and claims that one of the advantages of
using Guava is that it enables caches to be build at any time, rather
than only at server startup. Therefore, I guess it is OK to remove the
precondition check which prevents creation of caches once the server was
started.

This enables plugins to create own persistent caches when they are
installed.

Change-Id: I28b1a140c58f9967e6d1dacd61ddd021b48c8ad5

Former-commit-id: 218c6b69c288096630f5e17eec95689ea927b510